git-svn-id: http://gnuradio.org/svn/gnuradio/trunk@4541 221aa14e-8319-0410-a670-987f0aec2ac5

+++ gnuradio-core/src/lib/runtime/qa_gr_block.cc
@@ -42,7 +42,7 @@ qa_gr_block::t0 ()
   CPPUNIT_ASSERT_EQUAL (0, src1->input_signature()->max_streams ());
   CPPUNIT_ASSERT_EQUAL (1, src1->output_signature()->min_streams ());
   CPPUNIT_ASSERT_EQUAL (1, src1->output_signature()->max_streams ());
-  CPPUNIT_ASSERT_EQUAL (sizeof (int),
+  CPPUNIT_ASSERT_EQUAL ((int) sizeof(int),
 			src1->output_signature()->sizeof_stream_item (0));
 
   gr_block_sptr src2 (gr_make_null_source (sizeof (short)));
@@ -50,7 +50,7 @@ qa_gr_block::t0 ()
   CPPUNIT_ASSERT_EQUAL (0, src2->input_signature()->max_streams ());
   CPPUNIT_ASSERT_EQUAL (1, src2->output_signature()->min_streams ());
   CPPUNIT_ASSERT_EQUAL (1, src2->output_signature()->max_streams ());
-  CPPUNIT_ASSERT_EQUAL (sizeof (short),
+  CPPUNIT_ASSERT_EQUAL ((int) sizeof (short),
 			src2->output_signature()->sizeof_stream_item (0));
 }
 
@@ -63,7 +63,7 @@ qa_gr_block::t1 ()
   CPPUNIT_ASSERT_EQUAL (std::string ("null_sink"), dst1->name ());
   CPPUNIT_ASSERT_EQUAL (1, dst1->input_signature()->min_streams ());
   CPPUNIT_ASSERT_EQUAL (1, dst1->input_signature()->max_streams ());
-  CPPUNIT_ASSERT_EQUAL (sizeof (int),
+  CPPUNIT_ASSERT_EQUAL ((int) sizeof (int),
 			dst1->input_signature()->sizeof_stream_item (0));
 
   CPPUNIT_ASSERT_EQUAL (0, dst1->output_signature()->max_streams ());
@@ -72,7 +72,7 @@ qa_gr_block::t1 ()
   CPPUNIT_ASSERT_EQUAL (std::string ("null_sink"), dst2->name ());
   CPPUNIT_ASSERT_EQUAL (1, dst2->input_signature()->min_streams ());
   CPPUNIT_ASSERT_EQUAL (1, dst2->input_signature()->max_streams ());
-  CPPUNIT_ASSERT_EQUAL (sizeof (short),
+  CPPUNIT_ASSERT_EQUAL ((int) sizeof (short),
 			dst2->input_signature()->sizeof_stream_item (0));
   CPPUNIT_ASSERT_EQUAL (0, dst2->output_signature()->max_streams ());
 }
