+++ gnuradio-core/src/lib/runtime/qa_gr_hier_block2.cc
@@ -41,6 +41,6 @@ void qa_gr_hier_block2::test_make()
     CPPUNIT_ASSERT_EQUAL(1, src1->input_signature()->max_streams());
     CPPUNIT_ASSERT_EQUAL(1, src1->output_signature()->min_streams());
     CPPUNIT_ASSERT_EQUAL(1, src1->output_signature()->max_streams());
-    CPPUNIT_ASSERT_EQUAL(sizeof(int), 
+    CPPUNIT_ASSERT_EQUAL((int) sizeof(int), 
 			 src1->output_signature()->sizeof_stream_item(0));
 }
